+++ lib/flutter_billing.dart
@@ -7,6 +7,7 @@ export 'billing/billing_repository.dart';
 export 'billing/billing_request.dart';
 export 'billing/product_item.dart';
 export 'billing/purchase_response.dart';
+export 'billing/purchase_verifier.dart';
 export 'billing/signature_checker.dart';
 export 'di/package_di.dart';
 export 'presenter/billing_cubit.dart';